Further granularize the organization of bulk_invoke's implementation
