+++ agency/detail/bulk_functions/is_bulk_call_possible_via_execution_policy.hpp
